* merged part of 3904, 4223, 4225, 4231, 4234 & 4239 from trunk to 2.0 branch (airburst grenade ammo mode, new twohanded/onehanded system & code cleanups)

+++ src/client/keys.c
@@ -46,7 +46,7 @@ static int history_line = 0;
 
 int msg_mode;
 char msg_buffer[MAXCMDLINE];
-int msg_bufferlen = 0;
+size_t msg_bufferlen = 0;
 
 static int key_waiting;
 char *keybindings[256];
